refactor(crypto): extract transaction signing and utils out of the model
