+++ benchmark/crypto/hash-algorithms.js
@@ -7,7 +7,7 @@ const createHash = require("create-hash");
 const nodeSha256 = (bytes) => createHash("sha256").update(bytes).digest()
 
 const data = require('../helpers').getJSONFixture('transaction/deserialized/0');
-const transactionBytes = Transactions.Transaction.toBytes(data);
+const transactionBytes = Transactions.Utils.toBytes(data);
 
 exports['bcrypto.sha256'] = () => {
     HashAlgorithms.sha256(transactionBytes);
