+++ benchmark/transaction/serialize/0.js
@@ -5,5 +5,5 @@ const {
 const data = require('../../helpers').getJSONFixture('transaction/deserialized/0');
 
 exports['core'] = () => {
-    return Transactions.Transaction.toBytes(data);
+    return Transactions.Utils.toBytes(data);
 };
