+++ packages/crypto/src/transactions/factory.ts
@@ -5,7 +5,7 @@ import { BigNumber, isException } from "../utils";
 import { deserializer } from "./deserializer";
 import { Serializer } from "./serializer";
 import { TransactionTypeFactory } from "./types";
-import { Transaction } from "./types/transaction";
+import { Utils } from "./utils";
 import { Verifier } from "./verifier";
 
 export class TransactionFactory {
@@ -27,7 +27,7 @@ export class TransactionFactory {
     public static fromBytesUnsafe(buffer: Buffer, id?: string): ITransaction {
         try {
             const transaction = deserializer.deserialize(buffer);
-            transaction.data.id = id || Transaction.getId(transaction.data);
+            transaction.data.id = id || Utils.getId(transaction.data);
             transaction.isVerified = true;
 
             return transaction;
@@ -59,7 +59,7 @@ export class TransactionFactory {
 
         Serializer.serialize(transaction);
 
-        data.id = Transaction.getId(data);
+        data.id = Utils.getId(data);
         transaction.isVerified = transaction.verify();
 
         return transaction;
@@ -68,7 +68,7 @@ export class TransactionFactory {
     private static fromSerialized(serialized: string): ITransaction {
         try {
             const transaction = deserializer.deserialize(serialized);
-            transaction.data.id = Transaction.getId(transaction.data);
+            transaction.data.id = Utils.getId(transaction.data);
 
             const { value, error } = Verifier.verifySchema(transaction.data, true);
