+++ packages/crypto/src/transactions/index.ts
@@ -2,7 +2,9 @@ export * from "./builders";
 export * from "./deserializer";
 export * from "./factory";
 export * from "./serializer";
+export * from "./signer";
 export * from "./types";
+export * from "./utils";
 export * from "./verifier";
 
 export { transactionRegistry as TransactionRegistry, TransactionConstructor } from "./registry";
